Replace `KtLintMultiRule` (#6094)
* remove unused autoCorrect property

* WIP

* WIP

* fix documentation generation

* Allow multiple classes in rule set provider file

* remove multirule support in docs

* fix style and formatting violations

* revert removal of autocorrect prop

---------
