+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/collection/DetektCollector.kt
@@ -7,30 +7,25 @@ class DetektCollector : Collector<RuleSetPage> {
 
     private val ruleSetProviderCollector = RuleSetProviderCollector()
     private val ruleCollector = RuleCollector()
-    private val multiRuleCollector = MultiRuleCollector()
 
     private val collectors = listOf(
         ruleSetProviderCollector,
-        multiRuleCollector,
         ruleCollector
     )
     override val items: List<RuleSetPage>
         get() = buildRuleSetPages()
 
     private fun buildRuleSetPages(): List<RuleSetPage> {
-        val multiRules = multiRuleCollector.items
         val rules = ruleCollector.items
-        val multiRuleNameToRules = multiRules.associateBy({ it.name }, { it.rules })
         val ruleSets = ruleSetProviderCollector.items
 
         return ruleSets.map { ruleSet ->
-            val consolidatedRules = ruleSet.rules
-                .flatMap { ruleName -> multiRuleNameToRules[ruleName] ?: listOf(ruleName) }
+            val sortedRules = ruleSet.rules
                 .map { rules.findRuleByName(it) }
                 .sortedBy { rule -> rule.name }
 
-            consolidatedRules.resolveParentRule(rules)
-            RuleSetPage(ruleSet, consolidatedRules)
+            sortedRules.resolveParentRule(rules)
+            RuleSetPage(ruleSet, sortedRules)
         }
     }
