+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/collection/RuleSetProviderCollector.kt
@@ -9,10 +9,12 @@ import org.jetbrains.kotlin.psi.KtAnnotatedExpression
 import org.jetbrains.kotlin.psi.KtCallExpression
 import org.jetbrains.kotlin.psi.KtClass
 import org.jetbrains.kotlin.psi.KtFile
+import org.jetbrains.kotlin.psi.KtNameReferenceExpression
 import org.jetbrains.kotlin.psi.KtProperty
 import org.jetbrains.kotlin.psi.KtStringTemplateExpression
 import org.jetbrains.kotlin.psi.KtSuperTypeList
 import org.jetbrains.kotlin.psi.psiUtil.containingClass
+import org.jetbrains.kotlin.psi.psiUtil.findDescendantOfType
 import org.jetbrains.kotlin.psi.psiUtil.referenceExpression
 import io.gitlab.arturbosch.detekt.api.internal.Configuration as ConfigAnnotation
 
@@ -69,11 +71,13 @@ private class RuleSetProviderVisitor : DetektVisitor() {
     }
 
     override fun visitSuperTypeList(list: KtSuperTypeList) {
-        val superTypes = list.entries
-            ?.map { it.typeAsUserType?.referencedName }
-            ?.toSet()
-            .orEmpty()
-        containsRuleSetProvider = SUPPORTED_PROVIDERS.any { it in superTypes }
+        if (!containsRuleSetProvider) {
+            val superTypes = list.entries
+                ?.mapNotNull { it.typeAsUserType?.referencedName }
+                ?.toSet()
+                .orEmpty()
+            containsRuleSetProvider = SUPPORTED_PROVIDERS.any { it in superTypes }
+        }
         super.visitSuperTypeList(list)
     }
 
@@ -128,8 +132,10 @@ private class RuleSetProviderVisitor : DetektVisitor() {
 
         if (expression.calleeExpression?.text == "RuleSet") {
             val ruleListExpression = expression.valueArguments
-                .map { it.getArgumentExpression() }
-                .firstOrNull { it?.referenceExpression()?.text == "listOf" }
+                .mapNotNull { it.getArgumentExpression() }
+                .firstNotNullOfOrNull {
+                    it.findDescendantOfType<KtNameReferenceExpression> { exp -> exp.text == "listOf" }?.parent
+                }
                 ?: throw InvalidDocumentationException("RuleSetProvider $name doesn't provide list of rules.")
 
             val ruleArgumentNames = (ruleListExpression as? KtCallExpression)
